add register and forgort passwd function, some comment and rework password func for encryption. add try-catch block for login,reset and register function.

+++ EventHarbor/Screen/LoginScreen.xaml.cs
@@ -1,18 +1,8 @@
 ﻿using EventHarbor.Class;
 using EventHarbor.Screen;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace EventHarbor
 {
@@ -23,23 +13,29 @@ namespace EventHarbor
     {
         public LoginScreen()
         {
+            DatabaseContextManager dbManager = new DatabaseContextManager();
             InitializeComponent();
+
+            //temporary for development
+            if (dbManager.Database.EnsureDeleted())
+            {
+                MessageBox.Show("smazano");
+            }
+            if (dbManager.Database.EnsureCreated())
+            {
+                MessageBox.Show("Vytvořeno");
+            }
+
         }
 
-       
+
 
         private void CloseBtn_Click(object sender, RoutedEventArgs e)
         {
             Close();
-            
-        }
-
-
-        private void Hyperlink_Click_1(object sender, RoutedEventArgs e)
-        {
 
         }
-
+        //switch sreen to forgot window
         private void ForgotUri_Click(object sender, RoutedEventArgs e)
         {
             ForgotScreen screen = new ForgotScreen();
@@ -47,7 +43,7 @@ namespace EventHarbor
             screen.ShowDialog();
             this.Visibility = Visibility.Visible;
         }
-
+        //switch screen to register window
         private void RegisterUri_Click(object sender, RoutedEventArgs e)
         {
             RegisterScreen register = new RegisterScreen();
@@ -55,19 +51,45 @@ namespace EventHarbor
             register.ShowDialog();
             this.Visibility = Visibility.Visible;
         }
-
+        //move window function
         private void Border_MouseLeftButtonDown(object sender, MouseButtonEventArgs e)
         {
             var move = sender as Border;
             var win = Window.GetWindow(move);
             win.DragMove();
         }
-
+        //login function
         private void LoginBtn_Click(object sender, RoutedEventArgs e)
         {
 
-            using (DatabaseContextManager dbManager = new DatabaseContextManager())
+            UserManager userManager = new UserManager();
+
+            if (UserNameTextBox.Text != null && PwdBox.Password != null)
             {
+                try
+                {
+                    switch (userManager.IsRegistered(UserNameTextBox.Text, PwdBox.Password))
+                    {
+                        case -1:
+                            MessageBox.Show("Uživatel nenalezen");
+                            break;
+                        case 0:
+                            MessageBox.Show("Uživatel nalezen ale heslo je zadané chybně, prosím opakuj.");
+                            break;
+                        case 1:
+                            MessageBox.Show("Přihlášeno");
+                            break;
+
+                    }
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message);
+
+                }
+
+
+                /*
                 if (dbManager.Database.EnsureDeleted())
                 {
                     MessageBox.Show("smazano");
@@ -76,6 +98,8 @@ namespace EventHarbor
                 {
                     MessageBox.Show("Vytvořeno");
                 }
+                */
+
             }
         }
     }
